fix(winui): 修复 XAML 控件类型绑定与事件处理，完善 GitHub Actions 打包流水线

+++ src/MiAir.WinUI/ViewModels/DashboardViewModel.cs
@@ -51,7 +51,7 @@ public partial class DashboardViewModel : ObservableObject
     private string _tokenStatusText = "未登录小米账号";
 
     [ObservableProperty]
-    private int _volume = 50;
+    private double _volume = 50;
 
     public DashboardViewModel()
     {
